arch-review: extract SettingsAutomation.vue from SettingsView.vue
The automations tab (7,733-line SettingsView.vue, largest file in the repo)
was the read-mostly section sharing the least state with the other six tabs.
Extract it to web/src/components/settings/SettingsAutomation.vue following the
DeviceView.vue pattern: SettingsView stays the tab shell and mounts the child,
passing the shared automation data and the job-telemetry read helpers (which
the Models tab also renders) as props. The child owns the tab's render-only
state (expansion, trigger buttons) and the trigger/backfill actions.

SettingsView drops 630 lines (7,733 -> 7,103) and loses the dead .auto-* styles
that no template referenced. Automation markup, logic, and styles are byte-for-byte
identical apart from prop-boundary qualification (props.automationItems etc.).
mountSmoke (18 SettingsView assertions) passes; vue-tsc --noEmit passes.
